fix(eject): make run-task script CommonJS explicitly
the script run-task generated during eject generates CommonJS
unconditionally, which makes it fail if the project is running with type
module. By adding the file extension .cjs the script is working in both
CommonJS and ES Module projects. The additional eslint argument
'--format=cjs' while beeing the default was added to make the fact that
we rely on cjs here more explicit.